BayesianQuadrature

BayesianQuadrature.BayesModel

BayesianQuadrature.BayesQuad

BayesianQuadrature.PriorSampling

BayesianQuadrature.BayesModel — Type

BayesModel(prior, logintegrand)

Model inheriting from AbstractMCMC.AbstractModel. prior should be a multivariate distribution from Distributions.jl at the moment prior has to be a MvNormal but this will improved in a later version logintegrand should be the log of the function to integrate.

BayesianQuadrature.BayesQuad — Type

BayesQuad(k::Kernel; l=1.0, σ::Real=1.0)

Bayesian Quadrature object. You can pass any kernel and the lengthscale and variance will be extracted. l can be a Real, a AbstractVector or a LowerTriangular.

BayesianQuadrature.PriorSampling — Type

PriorSampling()

Sampler which will use the prior distribution from the given model to provide samples.