API

BayesianQuadrature.BayesModel

BayesianQuadrature.BayesQuad

BayesianQuadrature.PriorSampling

Models

BayesianQuadrature.BayesModel — Type

BayesModel(prior, logintegrand) <: AbstractBQModel

Model inheriting from AbstractMCMC.AbstractModel. prior should be a multivariate distribution from Distributions.jl at the moment prior has to be a MvNormal but this will improved in a later version logintegrand should be the log of the function to integrate.

Samplers

BayesianQuadrature.PriorSampling — Type

PriorSampling()

Sampler which will use the prior distribution from the given model to provide samples.

Bayesian Quadratures

BayesianQuadrature.BayesQuad — Type

BayesQuad(k::Kernel; l=1.0, σ::Real=1.0)

Tool for estimating the bayesian quadrature for a given BayesQuadModel and a Sampler.

BayesQuad estimate the probability distribution p(I) of I = ∫ f(x) p(x) dx

Argument

k::Kernel : kernel from KernelFunctions.jl, the kernel cannot be composite

Keywords argument

l : lengthscale of the kernel. It can be:

Real : isotropic kernel

AbstractVector : ARD kernel (one lengthscale per dimension)

LowerTriangular : Linear transformation of the inputs

σ : variance of the kernel (k(x,x') -> σ * k(x,x'))

Note: If k already has a variance and/or a transformation, these will be automatically extracted and replace the given keyword arguments