API

BayesianQuadrature.BayesQuad

BayesianQuadrature.PriorSampling

Models

Missing docstring.

Missing docstring for BayesianModel. Check Documenter's build log for details.

Samplers

BayesianQuadrature.PriorSampling — Type

PriorSampling()

Sampler which will use the prior distribution from the given model to provide samples.

Bayesian Quadratures

BayesianQuadrature.BayesQuad — Type

BayesQuad(k::Kernel; l=1.0, σ::Real=1.0)

Bayesian Quadrature object. You can pass any kernel and the lengthscale and variance will be extracted. l can be a Real, a AbstractVector or a LowerTriangular.